conroller-view add-show pizza

+++ src/main/java/fr/eni/PizzaOnlinenico/dal/PizzaDAO.java
@@ -1,11 +1,7 @@
 package fr.eni.PizzaOnlinenico.dal;
 
 import org.springframework.data.repository.CrudRepository;
-
 import fr.eni.PizzaOnlinenico.bo.Pizza;
 
-public interface PizzaDAO extends CrudRepository<Pizza, Integer>{
-	
-	
-
+public interface PizzaDAO extends CrudRepository<Pizza, Long>{
 }
